Conflicts:
	engine/src/main/java/org/teiid/dqp/internal/datamgr/ConnectorWorkItem.java
	engine/src/main/java/org/teiid/dqp/internal/process/DQPCore.java

+++ api/src/main/java/org/teiid/logging/CommandLogMessage.java
@@ -58,7 +58,7 @@ public class CommandLogMessage {
     private String requestID;
     private Long sourceCommandID;
     private String sql;
-    private Integer rowCount;
+    private Long rowCount;
     private String modelName;
     private String translatorName;
     private ExecutionContext executionContext;
@@ -89,7 +89,7 @@ public class CommandLogMessage {
                                 String principal,
                                 String vdbName,
                                 int vdbVersion, 
-                                Integer finalRowCount,
+                                Long finalRowCount,
                                 Event event, PlanNode plan) {
         // userCommandEnd
         this.event = event;
@@ -125,7 +125,7 @@ public class CommandLogMessage {
                                 String translatorName,
                                 String sessionID,
                                 String principal,
-                                Integer finalRowCount,
+                                Long finalRowCount,
                                 Event event,
                                 ExecutionContext context, Long cpuTime) {
         // dataSourceCommandEnd
@@ -195,7 +195,7 @@ public class CommandLogMessage {
 	 * Returns the command.  Only valid for {@link Event#END}
 	 * @return
 	 */
-	public Integer getRowCount() {
+	public Long getRowCount() {
 		return rowCount;
 	}
 	public String getModelName() {
